Refactor Exceptions into their own files (#229)
