+++ cogs/archive.py
@@ -11,7 +11,8 @@ from logging import Logger
 
 import discord
 
-from exceptions import BaseDoesNotExistError, DiscordMemberNotInMainGuildError
+from exceptions import DiscordMemberNotInMainGuildError
+from exceptions.base import BaseDoesNotExistError
 from utils import (
     CommandChecks,
     TeXBotApplicationContext,
