+++ utils/tex_bot_base_cog.py
@@ -15,10 +15,9 @@ from typing import TYPE_CHECKING, Final
 import discord
 from discord import Cog
 
-from exceptions import (
+from exceptions import CommitteeRoleDoesNotExistError, DiscordMemberNotInMainGuildError
+from exceptions.base import (
     BaseDoesNotExistError,
-    CommitteeRoleDoesNotExistError,
-    DiscordMemberNotInMainGuildError,
 )
 from utils.tex_bot import TeXBot
 from utils.tex_bot_contexts import TeXBotApplicationContext, TeXBotAutocompleteContext
